feature(updatePolicy): add an update policy
This commit adds a policy to enable an admin delete any thread however logged in users can delete their own threads

Delivers [#updatePolicy]

+++ app/Http/Controllers/ThreadsController.php
@@ -114,8 +114,12 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        // if ($thread->user_id != auth()->id()) {
 
-        $thread->replies()->delete();
+        //     abort(403, 'You do not have permission to do this');
+        // }
+
+        $this->authorize('update', $thread);
 
         $thread->delete();
